+++ lib/src/internal/sections/titled_widget.dart
@@ -1,42 +1,42 @@
+import 'package:demoflu/src/page.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
 /// Widget for title session.
 @internal
 class TitledWidget extends StatelessWidget {
-  const TitledWidget(
-      {super.key,
-      this.title,
-      required this.child,
-      required this.bordered,
-      required this.background,
-      required this.padding});
+  const TitledWidget({super.key, required this.section, required this.child});
 
-  final String? title;
-  final bool bordered;
+  final TitledPageSection section;
   final Widget child;
-  final Color? background;
-  final EdgeInsetsGeometry? padding;
 
   @override
   Widget build(BuildContext context) {
-    if (title != null) {
+    if (section.title != null) {
       return Column(
           crossAxisAlignment: CrossAxisAlignment.start,
-          children: [Text(title!), buildContent(context)]);
+          children: [Text(section.title!), buildContent(context)]);
     }
     return buildContent(context);
   }
 
   Widget buildContent(BuildContext context) {
-    if (bordered) {
-      return Container(
-          padding: padding,
-          decoration: BoxDecoration(
-              color: background,
-              border: Border.all(color: Colors.grey[200]!, width: 1)),
-          child: child);
-    }
-    return child;
+    Widget content = section.aspectRatio != null
+        ? AspectRatio(aspectRatio: section.aspectRatio!, child: child)
+        : child;
+    return ConstrainedBox(
+        constraints: BoxConstraints(
+            minWidth: section.minWidth,
+            minHeight: section.minHeight,
+            maxWidth: section.maxWidth,
+            maxHeight: section.maxHeight),
+        child: Container(
+            padding: section.padding,
+            decoration: BoxDecoration(
+                color: section.background,
+                border: section.bordered
+                    ? Border.all(color: Colors.grey[200]!, width: 1)
+                    : null),
+            child: content));
   }
 }